ai_domestic_accounts_book_配布用/src/requirements.txt
openai[azure]
python-dotenv
azure-ai-formrecognizer
azure-core
pillow
pillow-heif
matplotlib